+++ packages/base/form-element.ts
@@ -8,6 +8,8 @@
 // tslint:disable:strip-private-property-underscore
 
 
+import {property} from 'lit-element';
+
 import {addHasRemoveClass, BaseElement, CustomEventListener, EventType, SpecificEventListener} from './base-element';
 import {RippleInterface} from './utils';
 
@@ -17,8 +19,28 @@ export {
   CustomEventListener,
   EventType,
   RippleInterface,
-  SpecificEventListener,
+  SpecificEventListener
 };
+
+declare global {
+  interface FormDataEvent extends Event {
+    formData: FormData;
+  }
+
+  interface HTMLElementEventMap {
+    formdata: FormDataEvent;
+  }
+}
+
+// TODO(b/193921953): Remove when lit-html typing for ShadyDOM works internally
+interface WindowWithShadyDOM {
+  // tslint:disable-next-line:enforce-name-casing
+  ShadyDOM?: {inUse: boolean};
+}
+
+// ShadyDOM should submit <input> elements in component internals
+const USING_SHADY_DOM =
+    (window as unknown as WindowWithShadyDOM).ShadyDOM?.inUse ?? false;
 
 /** @soyCompatible */
 export abstract class FormElement extends BaseElement {
@@ -33,20 +55,64 @@ export abstract class FormElement extends BaseElement {
   protected abstract formElement: HTMLElement;
 
   /**
+   * Disabled state for the component. When `disabled` is set to `true`, the
+   * component will not be added to form submission.
+   */
+  @property({type: Boolean}) disabled = false;
+
+  /**
    * Implement ripple getter for Ripple integration with mwc-formfield
    */
   readonly ripple?: Promise<RippleInterface|null>;
 
-  click() {
-    if (this.formElement) {
-      this.formElement.focus();
-      this.formElement.click();
+  /**
+   * Form element that contains this element
+   */
+  protected containingForm: HTMLFormElement|null = null;
+  protected formDataListener = (ev: FormDataEvent) => {
+    if (!this.disabled) {
+      this.setFormData(ev.formData);
     }
+  };
+
+  protected findFormElement(): HTMLFormElement|null {
+    // If the component internals are not in Shadow DOM, subscribing to form
+    // data events could lead to duplicated data, which may not work correctly
+    // on the server side.
+    if (!this.shadowRoot || USING_SHADY_DOM) {
+      return null;
+    }
+    const root = this.getRootNode() as HTMLElement;
+    const forms = root.querySelectorAll('form');
+    for (const form of Array.from(forms)) {
+      if (form.contains(this)) {
+        return form;
+      }
+    }
+    return null;
   }
 
-  setAriaLabel(label: string) {
-    if (this.formElement) {
-      this.formElement.setAttribute('aria-label', label);
+  /**
+   * Implement this callback to submit form data
+   */
+  protected abstract setFormData(formData: FormData): void;
+
+  connectedCallback() {
+    super.connectedCallback();
+    this.containingForm = this.findFormElement();
+    this.containingForm?.addEventListener('formdata', this.formDataListener);
+  }
+
+  disconnectedCallback() {
+    super.disconnectedCallback();
+    this.containingForm?.removeEventListener('formdata', this.formDataListener);
+    this.containingForm = null;
+  }
+
+  click() {
+    if (this.formElement && !this.disabled) {
+      this.formElement.focus();
+      this.formElement.click();
     }
   }