+++ packages/select/mwc-select-base.ts
@@ -158,6 +158,8 @@ export abstract class SelectBase extends FormElement {
   })
   value = '';
 
+  @property() name = '';
+
   @state() protected selectedText = '';
 
   @property({type: String}) icon = '';
@@ -263,8 +265,10 @@ export abstract class SelectBase extends FormElement {
           class="mdc-select ${classMap(classes)}">
         <input
             class="formElement"
-            .value=${this.value}
+            name="${this.name}"
+            .value="${this.value}"
             hidden
+            ?disabled="${this.disabled}"
             ?required=${this.required}>
         <!-- @ts-ignore -->
         <div class="mdc-select__anchor"
@@ -865,6 +869,12 @@ export abstract class SelectBase extends FormElement {
     if (this.mdcFoundation) {
       this.menuOpen = false;
       this.mdcFoundation.handleMenuClosed();
+    }
+  }
+
+  protected setFormData(formData: FormData) {
+    if (this.name && this.selected !== null) {
+      formData.append(this.name, this.value);
     }
   }