+++ packages/textarea/test/mwc-textarea.test.ts
@@ -11,7 +11,7 @@
 import {TextArea} from '@material/mwc-textarea';
 import {html} from 'lit-html';
 
-import {fixture, TestFixture} from '../../../test/src/util/helpers';
+import {fixture, simulateFormDataEvent, TestFixture} from '../../../test/src/util/helpers';
 
 const basic = html`
   <mwc-textarea></mwc-textarea>
@@ -19,6 +19,12 @@ const basic = html`
 
 const withLabel = html`
   <mwc-textarea label="a label"></mwc-textarea>
+`;
+
+const textareaInForm = html`
+  <form>
+    <mwc-textarea name="foo"></mwc-textarea>
+  </form>
 `;
 
 describe('mwc-textarea:', () => {
@@ -77,4 +83,46 @@ describe('mwc-textarea:', () => {
       }
     });
   });
+
+  // IE11 can only append to FormData, not inspect it
+  if (Boolean(FormData.prototype.get)) {
+    describe('form submission', () => {
+      let form: HTMLFormElement;
+      let element: TextArea;
+
+      beforeEach(async () => {
+        fixt = await fixture(textareaInForm);
+        element = fixt.root.querySelector('mwc-textarea')!;
+        form = fixt.root.querySelector('form')!;
+        await element.updateComplete;
+      });
+
+      it('does not submit without a name', async () => {
+        element.name = '';
+        await element.updateComplete;
+        const formData = simulateFormDataEvent(form);
+        const keys = Array.from(formData.keys());
+        expect(keys.length).toEqual(0);
+      });
+
+      it('does not submit if disabled', async () => {
+        element.disabled = true;
+        await element.updateComplete;
+        const formData = simulateFormDataEvent(form);
+        expect(formData.get('foo')).toBeNull();
+      });
+
+      it('submits empty string by default', async () => {
+        const formData = simulateFormDataEvent(form);
+        expect(formData.get('foo')).toEqual('');
+      });
+
+      it('submits given value', async () => {
+        element.value = 'bar';
+        await element.updateComplete;
+        const formData = simulateFormDataEvent(form);
+        expect(formData.get('foo')).toEqual('bar');
+      });
+    });
+  }
 });